Refactor cart management and navigation in App and ShopPage components
